MangaController overhaul (#7244)
(cherry picked from commit 33a778873af0eb52528a29f741fa59b530679c64)

# Conflicts:
#	app/build.gradle.kts
#	app/src/main/java/eu/kanade/tachiyomi/data/database/models/LibraryManga.kt
#	app/src/main/java/eu/kanade/tachiyomi/ui/browse/migration/search/SearchController.kt
#	app/src/main/java/eu/kanade/tachiyomi/ui/manga/MangaController.kt
#	app/src/main/java/eu/kanade/tachiyomi/ui/manga/MangaPresenter.kt
#	app/src/main/java/eu/kanade/tachiyomi/ui/manga/chapter/ChapterHolder.kt
#	app/src/main/java/eu/kanade/tachiyomi/ui/manga/chapter/ChaptersAdapter.kt
#	app/src/main/java/eu/kanade/tachiyomi/ui/manga/chapter/ChaptersSettingsSheet.kt
#	app/src/main/java/eu/kanade/tachiyomi/ui/manga/info/MangaInfoHeaderAdapter.kt
#	app/src/main/java/eu/kanade/tachiyomi/widget/MangaSummaryView.kt
#	app/src/main/res/layout-sw720dp/manga_info_header.xml
#	app/src/main/res/layout/manga_controller.xml
#	app/src/main/res/layout/manga_info_header.xml
#	app/src/main/res/layout/manga_summary.xml
#	app/src/main/res/menu/manga.xml

+++ app/src/main/java/eu/kanade/tachiyomi/data/library/CustomMangaManager.kt
@@ -77,7 +77,7 @@ class CustomMangaManager(val context: Context) {
         val artist: String? = null,
         val description: String? = null,
         val genre: List<String>? = null,
-        val status: Int? = null,
+        val status: Long? = null,
     ) {
 
         fun toManga() = CustomMangaInfo(
@@ -87,7 +87,7 @@ class CustomMangaManager(val context: Context) {
             artist = this@MangaJson.artist,
             description = this@MangaJson.description,
             genre = this@MangaJson.genre,
-            status = this@MangaJson.status?.takeUnless { it == 0 },
+            status = this@MangaJson.status?.takeUnless { it == 0L },
         )
     }
 
@@ -98,7 +98,7 @@ class CustomMangaManager(val context: Context) {
         val artist: String? = null,
         val description: String? = null,
         val genre: List<String>? = null,
-        val status: Int? = null,
+        val status: Long? = null,
     ) {
         val genreString by lazy {
             genre?.joinToString()
